fix(pictor): Fable 全量診断の Critical/High 修正 (deadlock/heap破壊/font OOB/path traversal) (#76)
PR #75 の Fable 診断 (review/2026-06-11) で挙がった Critical 1 + High の
うち、明確に正しく外科的に直せる相関バグ群を一括修正。

- C-1 ThreadPoolDispatcher lost wakeup: worker の pending デクリメント後の
  notify を mutex 下に入れ、wait_all() の述語評価〜wait() 間での通知消失
  (毎フレーム踏むレンダースレッド永久ハング) を解消 (job_dispatcher.cpp)
- M-7 dispatch の pending_tasks_ 上書きを fetch_add に、M-8
  hardware_concurrency()==0 の uint32 ラップを guard (job/update_scheduler)
- H-2 FrameAllocator move 代入の解放関数ミスマッチ: large-pages ビルドで
  mmap/VirtualAlloc 領域を _aligned_free/free に渡すヒープ破壊を、dtor と
  共有する release_() に一本化して解消 (frame_allocator)
- Q-1 acquire 失敗時の fence デッドロック: fence reset を acquire 成功後
  (submit 前) に移動し、OUT_OF_DATE 時は signaled のまま残す (vulkan_context)
- V-1 TrueType glyf: total_pts を 32bit 計算 + ends 単調性検証で、末尾
  0xFFFF 巻き戻りと非単調 ends による heap OOB read を閉じる
- V-2 font テーブル/グリフオフセットの uint32 加算オーバーフローを size_t
  昇格で迂回不能に (text_rasterizer / font_loader)
- V-3 FileSystemResourceLoader: root_ 設定時は weakly_canonical で root 配下
  を検証し、絶対パス/.. トラバーサルを拒否 (resource_loader)
- H-1 BatchBuilder: static バッチを別キャッシュして毎フレームマージし、
  dirty が落ちた 2 フレーム目以降の静的オブジェクト消失を解消 (batch_builder)

検証: pictor.lib (Release, RIVE=OFF) ビルド成功、ctest 12/12 pass
(pixel_text_test / fps_baseline_test 含む)。残りの設計系 (D-1/D-2/D-3) と
DoS ハードニング (V-4〜V-10/H-3) は規模が大きく別 PR で対応予定。

+++ src/text/font_loader.cpp
@@ -222,7 +222,7 @@ std::vector<FontLoader::TableRecord> FontLoader::read_table_directory(
         // offsetTable[0] at byte 12
         if (size < 16) return records;
         uint32_t first_font_offset = read_u32(data + 12);
-        if (first_font_offset + 12 > size) return records;
+        if (static_cast<size_t>(first_font_offset) + 12 > size) return records;
         font_data = data + first_font_offset;
     }
 
@@ -267,7 +267,9 @@ bool FontLoader::parse_font_tables(FontTableEntry& entry) {
     const uint8_t* name_data = nullptr; uint32_t name_len = 0;
 
     for (const auto& rec : tables) {
-        if (rec.offset + rec.length > size) continue;
+        // size_t arithmetic: offset/length are file-controlled uint32s whose
+        // 32-bit sum could wrap and pass this check while pointing past the buffer.
+        if (static_cast<size_t>(rec.offset) + rec.length > size) continue;
         const uint8_t* tdata = data + rec.offset;
 
         if      (rec.tag == make_tag('h','e','a','d')) { head_data = tdata; head_len = rec.length; }
